rename modules to module
