+++ dspy/primitives/__init__.py
@@ -1,6 +1,6 @@
 from dspy.primitives.base_module import BaseModule
 from dspy.primitives.example import Example
-from dspy.primitives.modules import Module
+from dspy.primitives.module import Module
 from dspy.primitives.prediction import Completions, Prediction
 from dspy.primitives.python_interpreter import PythonInterpreter
