+++ dspy/teleprompt/bootstrap_finetune.py
@@ -12,7 +12,7 @@ from dspy.dsp.utils.settings import settings
 from dspy.evaluate.evaluate import Evaluate
 from dspy.predict.predict import Predict
 from dspy.primitives.example import Example
-from dspy.primitives.modules import Module
+from dspy.primitives.module import Module
 from dspy.teleprompt.teleprompt import Teleprompter
 from dspy.utils.exceptions import AdapterParseError
