+++ tests/primitives/test_module.py
@@ -1,6 +1,5 @@
-# TODO: Rename to test_module.py
 import dspy
-from dspy.primitives.modules import Module, set_attribute_by_name  # Adjust the import based on your file structure
+from dspy.primitives.module import Module, set_attribute_by_name  # Adjust the import based on your file structure
 from dspy.utils import DummyLM
